node: Move sqlx methods from utils to database mod
